+++ Numerical Methods/src/ui/events/InputListener.java
@@ -3,9 +3,9 @@ package ui.events;
 public interface InputListener {
     void equationChanged(String text);
     
-    void upperLimitChanged(int x);
-    void lowerLimitChanged(int x);
-    void startingPointChanged(int x);
+    void upperLimitChanged(double x);
+    void lowerLimitChanged(double x);
+    void startingPointChanged(double x);
     
     void iterationLimitChanged(int iterations);
     void iterationsSelected();
